WIP TODO decide where to record global-* ...
Right now we're doing it in the expander (and in $extract-source since
that had been optional).

Since we've moved recording lexical ref/set to $extract-source, that pass
is no longer optional when we're trying to get source info.

But that means the parameterize over in build-library-body don't
actually do anything (won't suppress add-global-set! for the set!
that install library-global exports).

So we may need another mechanism:
  1. "fix it in post" by clearing identifier-info-set-src* for exported id
  2. don't record the assignment by recognizing it as a library global;
     but we may or may not have a source-map loaded with the information
     we need
  3. leave the assignment and let clients deal with it
      - it'll have no source
      - problem is that it may then appear that the identifier is
        assigned somewhere such that a "rename all references" will
        refuse because there is some use that it can't find, namely
        the internal one